Заработало

+++ m16_architecture/app/src/main/java/com/popov/cleanarchitecture/data/UsefulActivityDto.kt
@@ -2,12 +2,12 @@ package com.popov.cleanarchitecture.data
 
 import com.popov.cleanarchitecture.entity.UsefulActivity
 
-class UsefulActivityDto(
+class UsefulActivityDto (
     override val accessibility: Double,
     override val activity: String,
     override val key: String,
     override val link: String,
-    override val participants: Int,
+    override val participants: Double,
     override val price: Double,
     override val type: String
 ) : UsefulActivity {
