+++ m16_architecture/app/src/main/java/com/popov/cleanarchitecture/entity/UsefulActivity.kt
@@ -5,7 +5,7 @@ interface UsefulActivity{
     val activity: String
     val key: String
     val link: String
-    val participants: Int
+    val participants: Double
     val price: Double
     val type: String
 }
